refactor(crate): split binary into lib.rs + main.rs following RUST.md layout
Move all module declarations, USER_AGENT, Cli, and the full run logic into
src/lib.rs. Reduce src/main.rs to a 5-line entry point that delegates to
`scout::run()`. Replace `exit()` calls with `ExitCode` return values throughout.
